requirements.txt
# Core dependencies
flask==3.0.0
werkzeug==3.0.1
requests==2.31.0
python-dateutil==2.8.2

# Scheduling
schedule==1.2.0

# Security
cryptography==41.0.7
pyjwt==2.8.0

# Image processing
Pillow==10.1.0
numpy==1.26.4
opencv-python==4.8.1.78

# Face recognition - critical order
face-recognition==1.3.0
face-recognition-models==0.3.0
dlib==19.24.4
cmake
